fix: build error store

+++ store/src/main/java/com/backend/store/dataobject/sqlentity/ArticleSqlEntity.java
@@ -44,7 +44,7 @@ public class ArticleSqlEntity extends ArticleAbs {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Long customerNumber = 0L;
-    private Long grade = null;
+    private Float grade = null;
 
     //specific
     public ArticleSqlEntity removeCategory(CategorySqlEntity categoryToRemove) {
